Chapter-4/saas_app/requirements.txt
fastapi[all]
sqlalchemy>=2.0.0
passlib
bcrypt==3.2.2
python-jose[cryptography]
httpx
pyotp
python-dotenv

# testing packages
pytest
